# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Social App API
  description: API documentation for the Social App backend
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /api/bulletin/{id}:
    put:
      tags:
      - Bulletin Posts
      summary: Update a post
      description: Updates an existing post by its ID
      operationId: updatePosts
      parameters:
      - name: id
        in: path
        description: ID of the post to update
        required: true
        schema:
          type: integer
          format: int32
        example: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Posts"
        required: true
      responses:
        "200":
          description: Post updated successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Posts"
        "404":
          description: Post not found
        "400":
          description: Invalid post data
      security:
      - BearerAuth: []
...
  /api/bulletin/posts:
    get:
      tags:
      - Bulletin Posts
      summary: Get bulletin feed
      description: Returns the bulletin board feed for the caller's neighbourhood
        zone with optional filtering and pagination
      operationId: getAllPosts
      parameters:
      - name: category
        in: query
        description: Optional category filter
        required: false
        schema:
          type: string
        example: General
      - name: search
        in: query
        description: Optional keyword search against post content
        required: false
        schema:
          type: string
        example: help
      - name: page
        in: query
        description: Page number for pagination
        required: false
        schema:
          type: integer
          format: int32
        example: 1
      - name: limit
        in: query
        description: Number of posts per page
        required: false
        schema:
          type: integer
          format: int32
        example: 20
      - name: Authorization
        in: header
        description: "Firebase authentication token in format: 'Bearer <token>'"
        required: true
        schema:
          type: string
        example: Bearer eyJhbGciOiJSUzI1NiIsImtpZCI6...
      responses:
        "200":
          description: Feed retrieved successfully
          content:
            '*/*':
              schema:
                type: object
        "401":
          description: Invalid or expired Firebase token
      security:
      - BearerAuth: []
    post:
      tags:
      - Bulletin Posts
      summary: Create a new post
      description: Creates a new bulletin post by an authenticated user
      operationId: createPosts
      parameters:
      - name: Authorization
        in: header
        description: "Firebase authentication token in format: 'Bearer <token>'"
        required: true
        schema:
          type: string
        example: Bearer eyJhbGciOiJSUzI1NiIsImtpZCI6...
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostRequest"
        required: true
      responses:
        "201":
          description: Post created successfully
          content:
            '*/*':
              schema:
                type: object
        "401":
          description: Invalid or expired Firebase token
        "400":
          description: Invalid post data
      security:
      - BearerAuth: []
...
  /api/bulletin/posts/{postId}:
    get:
      tags:
      - Bulletin Posts
      summary: Get post details
      description: Returns a single post's detail view including its full comments
        list
      operationId: getPostsById
      parameters:
      - name: postId
        in: path
        description: ID of the post to retrieve
        required: true
        schema:
          type: integer
          format: int32
        example: 1
      - name: Authorization
        in: header
        description: "Firebase authentication token in format: 'Bearer <token>'"
        required: true
        schema:
          type: string
        example: Bearer eyJhbGciOiJSUzI1NiIsImtpZCI6...
      responses:
        "200":
          description: Post details retrieved successfully
          content:
            '*/*':
              schema:
                type: object
        "401":
          description: Invalid or expired Firebase token
        "404":
          description: Post not found
      security:
      - BearerAuth: []
    delete:
      tags:
      - Bulletin Posts
      summary: Delete a post
      description: Deletes a post by its ID if the authenticated user is the owner
      operationId: deletePosts
      parameters:
      - name: postId
        in: path
        description: ID of the post to delete
        required: true
        schema:
          type: integer
          format: int32
        example: 1
      - name: Authorization
        in: header
        description: "Firebase authentication token in format: 'Bearer <token>'"
        required: true
        schema:
          type: string
        example: Bearer eyJhbGciOiJSUzI1NiIsImtpZCI6...
      responses:
        "200":
          description: Post deleted successfully
          content:
            '*/*':
              schema:
                type: object
        "401":
          description: Invalid or expired Firebase token
        "403":
          description: Not authorized to delete this post
        "404":
          description: Post not found
      security:
      - BearerAuth: []
  /api/bulletin/posts/{postId}/comments:
    get:
      tags:
      - Bulletin Posts
      summary: Get comments for a post
      description: Returns all comments for a specific post
      operationId: getPostComments
      parameters:
      - name: postId
        in: path
        description: ID of the post to get comments for
        required: true
        schema:
          type: integer
          format: int32
        example: 1
      - name: Authorization
        in: header
        description: "Firebase authentication token in format: 'Bearer <token>'"
        required: true
        schema:
          type: string
        example: Bearer eyJhbGciOiJSUzI1NiIsImtpZCI6...
      responses:
        "200":
          description: Comments retrieved successfully
          content:
            '*/*':
              schema:
                type: object
        "401":
          description: Invalid or expired Firebase token
        "404":
          description: Post not found
      security:
      - BearerAuth: []
...
  /api/comments/bulletin/posts/{postId}/{commentId}:
    delete:
      tags:
      - Comments
      summary: Delete a comment from a post
      description: Deletes a comment from a specific bulletin post
      operationId: deleteCommentsUnderPost_1
      parameters:
      - name: postId
        in: path
        description: ID of the bulletin post
        required: true
        schema:
          type: integer
          format: int32
        example: 1
      - name: commentId
        in: path
        description: ID of the comment to delete
        required: true
        schema:
          type: integer
          format: int32
        example: 1
      - name: Authorization
        in: header
        description: "Firebase authentication token in format: 'Bearer <token>'"
        required: true
        schema:
          type: string
        example: Bearer eyJhbGciOiJSUzI1NiIsImtpZCI6...
      responses:
        "204":
          description: Comment deleted successfully
        "401":
          description: Invalid or expired Firebase token
        "404":
          description: Comment or post not found
components:
  schemas:
    Address:
      type: object
      properties:
        addressid:
          type: integer
          format: int32
        streetNumber:
          type: integer
          format: int32
        street:
          type: string
        zipcode:
          type: integer
          format: int32
        neighbourhoodid:
          $ref: "#/components/schemas/Location"
...
    Badges:
      type: object
      properties:
        badgeid:
          type: integer
          format: int32
        badgeName:
          type: string
        badge_description:
          type: string
        isSpecialist:
          type: boolean
        xpReward:
          type: integer
          format: int32
        ratingid:
          $ref: "#/components/schemas/Ratings"
        badgeDescription:
          type: string
...
    Location:
      type: object
      properties:
        locationid:
          type: integer
          format: int32
        locationCenterPoint:
          type: integer
          format: int32
        locationRadius:
          type: integer
          format: int32
        neighbourhoodid:
          type: integer
          format: int32
        neighbourhoodName:
          type: string
    Ratings:
      type: object
      properties:
        ratingid:
          type: integer
          format: int32
        ratingReview:
          type: string
        totalXpLevel:
          type: integer
          format: int32
        currentGroup:
          type: string
...
    User:
      type: object
      properties:
        userid:
          type: integer
          format: int32
        firebaseUid:
          type: string
        emailVerified:
          type: boolean
        phoneVerified:
          type: boolean
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phoneNumber:
          type: string
        dateOfBirth:
          type: string
          format: date-time
        gender:
          type: string
        addressid:
          $ref: "#/components/schemas/Address"
        badgeid:
          $ref: "#/components/schemas/Badges"
        ratingid:
          $ref: "#/components/schemas/Ratings"
        userType:
          type: string
        isAdmin:
          type: boolean
...
    Posts:
      type: object
      properties:
        postid:
          type: integer
          format: int32
        userid:
          $ref: "#/components/schemas/User"
        postContent:
          type: string
        mediaURL:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        category:
          type: string
...
    CreatePostRequest:
      type: object
      properties:
        postContent:
          type: string
        mediaUrl:
          type: string
        category:
          type: string
...